Add crossing function. Haven't tested yet
